refactor: handlers for card
