+++ src/app/components/main.ts
@@ -39,7 +39,7 @@ import {
 
 import { addFilterNavigationForCheckbox, addFilterNavigationForPrices, addFilterNavigationForSearch, addFilterNavigationForSelect, addNavigationForSidebar, clearFilterForPrices, clearFilterForSearch, redrawProducts, toggleCatalogAddProductButton } from './handlers-catalog';
 import { changeCartItemQuantityFromCart } from './handlers-cart';
-import { toggleCardAddProductButton, changeCartItemQuantityFromCard } from './handlers-card';
+import { toggleCardAddProductButton, changeCartItemQuantityFromCard, switchNextSlide, switchPrevSlide, getImagebyMini, toggleCardModal } from './handlers-card';
 import Header from './header';
 
 class Main {
@@ -123,107 +123,6 @@ class Main {
     }
   }
 
-  private makeMiniActive(slideIndex: number): void {
-    const minis = document.querySelectorAll('.product-card__mini') as NodeListOf<HTMLDivElement>;
-
-    minis.forEach((mini) => {
-      mini.classList.remove('active-mini');
-    });
-    minis[slideIndex - 1].classList.add('active-mini');
-  }
-
-  private switchNextSlide(): void {
-    const cardWrapper = document.querySelector('.product-card') as HTMLDivElement;
-    let currentIndex = +cardWrapper.getAttribute('data-slideIndex')!;
-
-    const slidesRow = document.querySelector('.product-card__slides-row') as HTMLDivElement;
-    const slides = document.querySelectorAll('.product-card__slide') as NodeListOf<HTMLDivElement>;
-    const width = slides[0].clientWidth + 60;
-
-    const slidesRowModal = document.querySelector('.modal-card__slides-row') as HTMLDivElement;
-    const slidesModal = document.querySelectorAll('.modal-card__slide') as NodeListOf<HTMLDivElement>;
-    const widthModal = slidesModal[0].clientWidth + 60;
-
-    if (currentIndex > slides.length - 1) {
-      currentIndex = 1;
-      cardWrapper.setAttribute('data-slideIndex', `${currentIndex}`);
-
-      slidesRow.style.transform = `none`;
-      slidesRowModal.style.transform = `none`;
-    } else {
-      slidesRow.style.transform = `translateX(${-width * currentIndex}px)`;
-      slidesRowModal.style.transform = `translateX(${-widthModal * currentIndex}px)`;
-
-      currentIndex++;
-      cardWrapper.setAttribute('data-slideIndex', `${currentIndex}`);
-    }
-    this.makeMiniActive(currentIndex);
-  }
-
-  private switchPrevSlide(): void {
-    const cardWrapper = document.querySelector('.product-card') as HTMLDivElement;
-    let currentIndex = +cardWrapper.getAttribute('data-slideIndex')!;
-
-    const slidesRow = document.querySelector('.product-card__slides-row') as HTMLDivElement;
-    const slides = document.querySelectorAll('.product-card__slide') as NodeListOf<HTMLDivElement>;
-    const width = slides[0].clientWidth + 60;
-
-    const slidesRowModal = document.querySelector('.modal-card__slides-row') as HTMLDivElement;
-    const slidesModal = document.querySelectorAll('.modal-card__slide') as NodeListOf<HTMLDivElement>;
-    const widthModal = slidesModal[0].clientWidth + 60;
-
-    if (currentIndex <= 1) {
-      currentIndex = slides.length;
-      cardWrapper.setAttribute('data-slideIndex', `${currentIndex}`);
-
-      slidesRow.style.transform = `translateX(${-width * (currentIndex - 1)}px)`;
-      slidesRowModal.style.transform = `translateX(${-widthModal * (currentIndex - 1)}px)`;
-    } else {
-      currentIndex--;
-      cardWrapper.setAttribute('data-slideIndex', `${currentIndex}`);
-
-      slidesRow.style.transform = `translateX(${-width * (currentIndex - 1)}px)`;
-      slidesRowModal.style.transform = `translateX(${-widthModal * (currentIndex - 1)}px)`;
-    }
-    this.makeMiniActive(currentIndex);
-  }
-
-  private getImagebyMini(target: HTMLElement): void {
-    const cardWrapper = document.querySelector('.product-card') as HTMLDivElement;
-    let currentIndex = +cardWrapper.getAttribute('data-slideIndex')!;
-
-    const slidesRow = document.querySelector('.product-card__slides-row') as HTMLDivElement;
-    const slides = document.querySelectorAll('.product-card__slide') as NodeListOf<HTMLDivElement>;
-    const width = slides[0].clientWidth + 60;
-
-    currentIndex = +target.getAttribute('data-index')!;
-    cardWrapper.setAttribute('data-slideIndex', `${currentIndex}`);
-    slidesRow.style.transform = `translateX(${-width * (currentIndex - 1)}px)`;
-    this.makeMiniActive(currentIndex);
-  }
-
-  private toggleCardModal(slideIndex: number): void {
-    const modal = document.querySelector('.modal-card__dimming');
-    modal?.classList.toggle('modal-active');
-    if (modal?.classList.contains('modal-active')) {
-      const slidesRow = document.querySelector('.modal-card__slides-row') as HTMLDivElement;
-      const slides = document.querySelectorAll('.modal-card__slide') as NodeListOf<HTMLDivElement>;
-      const width = slides[0].clientWidth + 60;
-
-      slidesRow.style.transform = `translateX(${-width * (slideIndex - 1)}px)`;
-    } else {
-      const slidesRow = document.querySelector('.product-card__slides-row') as HTMLDivElement;
-      const slides = document.querySelectorAll('.product-card__slide') as NodeListOf<HTMLDivElement>;
-      const miniRow = document.querySelector('.product-card__minis-row');
-      const width = slides[0].clientWidth + 60;
-
-      slidesRow.style.transform = `translateX(${-width * (slideIndex - 1)}px)`;
-      if (miniRow) {
-        this.makeMiniActive(slideIndex);
-      }
-    }
-  }
-
   private setEventListeners(router: Router): void {
     document.addEventListener('click', async (event: Event): Promise<void> => {
       const target = event.target as HTMLElement;
@@ -303,18 +202,18 @@ class Main {
         target.classList.contains('product-card__next-slide') ||
         target.classList.contains('modal-card__next-slide')
       ) {
-        this.switchNextSlide();
+        switchNextSlide();
       }
 
       if (
         target.classList.contains('product-card__prev-slide') ||
         target.classList.contains('modal-card__prev-slide')
       ) {
-        this.switchPrevSlide();
+        switchPrevSlide();
       }
 
       if (target.classList.contains('product-card__mini-img')) {
-        this.getImagebyMini(target);
+        getImagebyMini(target);
       }
 
       if (
@@ -324,7 +223,7 @@ class Main {
       ) {
         const cardWrapper = document.querySelector('.product-card') as HTMLDivElement;
         const currentIndex = +cardWrapper.getAttribute('data-slideIndex')!;
-        this.toggleCardModal(currentIndex);
+        toggleCardModal(currentIndex);
       }
 
       if (target.classList.contains('product-card__add-to-cart')) {
